+++ README.md
@@ -3,8 +3,8 @@
   ![Level+ logo](https://github.com/NathanielSunday/LevelPlus/raw/main/icon.png)
 
   ![Version](https://img.shields.io/github/v/release/nathanielsunday/levelplus?color=green&label=latest%20version)
-  ![Release date](https://img.shields.io/steam/release-date/2600460955?label=first%20release)\
-  [![Steam users](https://img.shields.io/steam/downloads/2600460955?color=005777&logo=steam&label=steam%20users)](https://steamcommunity.com/sharedfiles/filedetails/?id=2600460955)
+  ![Release date](https://img.shields.io/steam/release-date/2600460955?label=steam%20release)\
+  [![Steam users](https://img.shields.io/steam/downloads/2600460955?color=005777&logo=steam)](https://steamcommunity.com/sharedfiles/filedetails/?id=2600460955)
   ![Time spent since 2025](https://wakatime.com/badge/user/018c82bf-577a-4897-9034-fbada0a44012/project/018c82e9-a0aa-497d-a64e-443c349e3c63.svg)\
   [![Discord](https://img.shields.io/discord/1298577898285961258?color=5865F2&logo=discord&label=discord)](https://discord.gg/tgHpw4GP45)
   [![Support me on Patreon](https://img.shields.io/badge/dynamic/json?url=https%3A%2F%2Fwww.patreon.com%2Fapi%2Fcampaigns%2F8927399&query=data.attributes.patron_count&suffix=%20patrons&label=patreon&color=orangered&logo=patreon)](https://patreon.com/poctordepper5446)
